chore: refactor outbox mod
