+++ crates/precompiles/src/lib.rs
@@ -28,19 +28,19 @@ pub(crate) use tempo_precompiles::{error, storage};
 pub mod aes_gcm;
 pub mod chaum_pedersen;
 pub mod ecies;
+pub mod outbox;
 pub mod policy;
 pub mod tempo_state;
 pub mod tip20_factory;
 pub mod tip403_proxy;
-pub mod zone_outbox;
 pub mod ztip20;
 
 pub use aes_gcm::{AES_GCM_DECRYPT_ADDRESS, AesGcmDecrypt};
 pub use chaum_pedersen::{CHAUM_PEDERSEN_VERIFY_ADDRESS, ChaumPedersenVerify};
+pub use outbox::{ZoneOutbox, ZonePortalReader};
 pub use tempo_state::{L1StorageReader, TempoState};
 pub use tip20_factory::{ZONE_TIP20_FACTORY_ADDRESS, ZoneTokenFactory};
 pub use tip403_proxy::{ZONE_TIP403_PROXY_ADDRESS, ZoneTip403ProxyRegistry};
-pub use zone_outbox::{ZoneOutbox, ZonePortalReader};
 pub use ztip20::{SequencerExt, ZoneTip20Token};
 
 use revm::precompile::PrecompileError;
